Move ServerCapabilities to Client directory
