+++ src/Client/GRPC/ServiceClient.php
@@ -10,7 +10,7 @@
 namespace Temporal\Client\GRPC;
 
 use Temporal\Api\Workflowservice\V1;
-use Temporal\Client\DTO\ServerCapabilities;
+use Temporal\Client\ServerCapabilities;
 use Temporal\Exception\Client\ServiceClientException;
 
 /**
